Move data_fabfile.py into cloudbio package so it can be used a library, create a symbolic link to this new location at the top-level to support current uses of this script.
